@@ -18,33 +18,38 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.9.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   provider: ^4.3.2+2
 
+  uuid: ^3.0.4
   
+  # AWS dependencies 
+  amplify_flutter: ^0.2.0
+  amplify_auth_cognito: ^0.2.0
+  http: ^0.13.0
   # Firebase Dependencies
-  firebase_core: 0.5.0+1
+  # firebase_core: 0.5.0+1
 
-  firebase_auth: ^0.18.1+2
-  firebase_auth_mocks: ^0.3.2
+  # firebase_auth: ^0.18.1+2
+  # firebase_auth_mocks: ^0.3.2
 
-  firebase_storage: ^4.0.1
-  firebase_storage_mocks: ^0.2.1
+  # firebase_storage: ^4.0.1
+  # firebase_storage_mocks: ^0.2.1
 
-  cloud_firestore: ^0.14.1+3
-  cloud_firestore_mocks: ^0.5.0+1
+  # cloud_firestore: ^0.14.1+3
+  # cloud_firestore_mocks: ^0.5.0+1
 
-  image_picker: ^0.6.7+12
-  path_provider: ^1.6.21
+  image_picker: ^0.8.4
+  path_provider: ^2.0.3
 
-  mockito: ^4.1.2
+  mockito: ^5.0.15
   get_it: ^3.1.0
 
-  google_fonts: ^1.1.1
+  google_fonts: ^2.1.0
   intl: ^0.16.1
 
   flutter_circular_slider: ^2.5.0
