@@ -1,5 +1,9 @@
 import 'package:fitnet/models/workout.dart';
 import 'package:fitnet/models/workoutStepInstance.dart';
+import 'package:fitnet/services/workoutInstanceService.dart';
+import 'package:uuid/uuid.dart';
+
+import '../serviceInjector.dart';
 
 class WorkoutInstance {
   String iid;
@@ -8,14 +12,20 @@ class WorkoutInstance {
   DateTime start;
   DateTime end;
   List<WorkoutStepInstance> steps = [];
+  WorkoutInstanceService service = injector<WorkoutInstanceService>();
+  Uuid uuid = injector<Uuid>();
+  
 
   WorkoutInstance.fromWorkout(Workout workout) {
+    iid = uuid.v4();
     wid = workout.wid;
     uid = workout.uid;
     steps = workout.steps
         .map((step) =>
             WorkoutStepInstance.forStep(step.formatType, step))
         .toList();
+    start = DateTime.now();
+    service.addInstance(this);
   }
 
   WorkoutInstance.fromMap(Map<String, dynamic> map) {
@@ -23,9 +33,9 @@ class WorkoutInstance {
     wid = map['wid'];
     uid = map['uid'];
     if (map['start'] != null)
-      start = DateTime.fromMillisecondsSinceEpoch(map['start'].seconds * 1000);
+      start = DateTime.fromMillisecondsSinceEpoch(map['start']);
     if (map['end'] != null)
-      end = DateTime.fromMillisecondsSinceEpoch(map['end'].seconds * 1000);
+      end = DateTime.fromMillisecondsSinceEpoch(map['end']);
     if (map['steps'] != null) {
       List<WorkoutStepInstance> mappedSteps = [];
       map['steps'].forEach((step) => mappedSteps
@@ -36,23 +46,43 @@ class WorkoutInstance {
 
   Map<String, dynamic> toMap() {
     Map<String, dynamic> map = {
+      'iid': iid,
       'wid': wid,
       'uid': uid,
     };
-    if (start != null) map['start'] = start;
-    if (end != null) map['end'] = end;
+    if (start != null) map['start'] = start.millisecondsSinceEpoch;
+    if (end != null) map['end'] = end.millisecondsSinceEpoch;
     if (steps != null)
       map['steps'] = steps.map((step) => step.toMap()).toList();
     return map;
   }
 
-  isCompleted() {
-    return this.end != null;
+  complete() {
+    end = DateTime.now();
+    steps.firstWhere((step) => !step.isCompleted, orElse: () => null)?.complete();
+    _update();
   }
 
-  double percentComplete() {
+  updateStep(WorkoutStepInstance step, [num index]) {
+    steps[index ?? currentStepIndex] = step;
+    _update();
+  }
+
+  _update() {
+    service.updateInstance(this);
+  }
+
+  bool get isStarted => this.start != null;
+
+  bool get isCompleted => this.end != null;
+
+  WorkoutStepInstance get currentStep => this.steps.firstWhere((step) => !step.isCompleted, orElse: () => null);
+
+  int get currentStepIndex => this.steps.indexWhere((step) => !step.isCompleted);
+
+  double get percentComplete {
     var percentage = 0.0;
-    steps.forEach((step) => percentage += step.percentComplete());
+    steps.forEach((step) => percentage += step.percentComplete);
     return (percentage / steps.length * 10).round() / 10;
   }
 }
