@@ -1,24 +1,20 @@
-import 'package:fitnet/services/authService.dart';
 import 'package:fitnet/routes/auth/authScreen.dart';
+import 'package:fitnet/routes/authChangeNotifier.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:provider/provider.dart';
 
-import '../serviceinjector.dart';
 import 'home/homeScreen.dart';
 
 class AuthRouter extends StatelessWidget {
-  final AuthService authService = injector<AuthService>();
 
   @override
   Widget build(BuildContext context) {
-    return StreamProvider(
-        create: (_) => authService
-            .getLoggedInUser()
-            .map((user) => user != null ? user.uid : null),
-        child: Consumer<String>(
-            builder: (_, uid, __) => getScreenForAuthStatus(uid)));
+    return ChangeNotifierProvider(
+        create: (_) => AuthChangeNotifier(isConfirmed: false),
+        child: Consumer<AuthChangeNotifier>(
+            builder: (_, notifier, __) => getScreenForAuthStatus(notifier)));
   }
 
-  Widget getScreenForAuthStatus(String uid) =>
-      uid != null ? HomeScreen(userId: uid) : AuthScreen();
+  Widget getScreenForAuthStatus(AuthChangeNotifier notifier) =>
+    !notifier.isConfirmed ? AuthScreen() : HomeScreen();
 }
