@@ -1,11 +1,13 @@
 import 'package:fitnet/models/workout.dart';
 import 'package:fitnet/routes/editWorkout/editWorkoutScreen.dart';
+import 'package:fitnet/routes/home/workout/workoutPageChangeNotifier.dart';
 import 'package:fitnet/routes/home/workout/workoutListItem/tagsDisplay/tagsDisplayRow.dart';
 import 'package:fitnet/routes/viewWorkout/viewWorkoutScreen.dart';
 import 'package:fitnet/routes/workoutHistory/workoutHistoryScreen.dart';
 import 'package:fitnet/services/workoutService.dart';
 import 'package:fitnet/utils/customColors.dart';
 import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
 
 import '../../../../serviceInjector.dart';
 
@@ -93,21 +95,30 @@ class WorkoutListItem extends StatelessWidget {
         break;
       case 'Duplicate':
         workout.wid = null;
-        workoutService.addOrUpdateWorkout(workout);
+        Workout dup = await workoutService.addOrUpdateWorkout(workout);
+        Provider.of<WorkoutPageChangeNotifier>(context, listen: false).addWorkout(dup);
         break;
       case 'History':
         Navigator.of(context).push(MaterialPageRoute(
             builder: (context) => WorkoutHistoryScreen(workout: workout)));
         break;
       case 'Delete':
-        workoutService.deleteWorkout(workout);
+        await workoutService.deleteWorkout(workout);
+        Provider.of<WorkoutPageChangeNotifier>(context, listen: false).removeWorkout(workout);
+
         break;
     }
   }
 
-  showEditWorkoutScreen(BuildContext context) {
-    Navigator.of(context).push(MaterialPageRoute(
-        builder: (context) => EditWorkoutScreen(workout: workout)));
+  showEditWorkoutScreen(BuildContext context) async  {
+    Workout editedWorkout = await Navigator.of(context).push(
+      MaterialPageRoute(
+        builder: (context) => EditWorkoutScreen(
+          workout: workout
+        ),
+      ),
+    );
+    Provider.of<WorkoutPageChangeNotifier>(context, listen: false).updateWorkout(editedWorkout);
   }
 
   showViewWorkoutScreen(BuildContext context) {
