@@ -1,54 +1,56 @@
 import 'package:circular_countdown_timer/circular_countdown_timer.dart';
 import 'package:fitnet/models/workoutStepInstance.dart';
+import 'package:fitnet/shared/notifiers/amrapChangeNotifier.dart';
+import 'package:fitnet/shared/notifiers/instanceChangeNotifier.dart';
 import 'package:fitnet/utils/customColors.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_circular_slider/flutter_circular_slider.dart';
 import 'package:provider/provider.dart';
 
-import '../../../shared/notifiers/amrapChangeNotifier.dart';
+import '../../../shared/mixins/timerControlMixin.dart';
 
 class AMRAPStepScreen extends StatelessWidget {
-  final AMRAPStepInstance step;
-  final Function nextStep;
 
-  AMRAPChangeNotifier _notifier;
   CountDownController _controller;
 
-  AMRAPStepScreen({@required this.step, @required this.nextStep}) {
-    _notifier = AMRAPChangeNotifier(step: step);
+  AMRAPStepScreen() {
     _controller = CountDownController();
   }
 
   @override
   Widget build(BuildContext context) {
-    return ChangeNotifierProvider<AMRAPChangeNotifier>.value(
-      value: _notifier,
-      builder: (_, __) => Consumer<AMRAPChangeNotifier>(
-        builder: (ctx, notifier, __) => Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Stack(
-              alignment: AlignmentDirectional.center,
-              children: [
-                Container(
-                  height: 385,
-                  width: 385,
-                  decoration: BoxDecoration(
-                      shape: BoxShape.circle,
-                      color: Colors.white,
-                      boxShadow: [
-                        BoxShadow(
-                            offset: Offset(3, 5),
-                            blurRadius: 9,
-                            color: CustomColors.darkGrey)
-                      ]),
-                ),
-                SingleCircularSlider(
-                  notifier.step.targetReps,
-                  notifier.step.actualReps % notifier.step.targetReps,
+    return ChangeNotifierProxyProvider<InstanceChangeNotifier, AMRAPChangeNotifer>(
+      create: (ctx) => AMRAPChangeNotifer(parent: Provider.of<InstanceChangeNotifier>(ctx, listen: false)),
+      update: (_, instance, notifier) {
+        notifier.setParent(instance);
+        return notifier;
+      },
+      builder: (ctx, __) => Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Stack(
+            alignment: AlignmentDirectional.center,
+            children: [
+              Container(
+                height: 385,
+                width: 385,
+                decoration: BoxDecoration(
+                    shape: BoxShape.circle,
+                    color: Colors.white,
+                    boxShadow: [
+                      BoxShadow(
+                          offset: Offset(3, 5),
+                          blurRadius: 9,
+                          color: CustomColors.darkGrey)
+                    ]),
+              ),
+              Consumer<AMRAPChangeNotifer>(
+                builder: (_, notifier, __) => SingleCircularSlider(
+                  notifier.currentStep.targetReps,
+                  notifier.currentStep.actualReps % notifier.currentStep.targetReps,
                   shouldCountLaps: true,
                   onSelectionChange: (_, value, laps) =>
-                      notifier.setReps(value + notifier.step.targetReps * laps),
+                      notifier.setReps(value + notifier.currentStep.targetReps * laps),
                   height: 400,
                   width: 400,
                   selectionColor: CustomColors.blue,
@@ -58,8 +60,9 @@ class AMRAPStepScreen extends StatelessWidget {
                     crossAxisAlignment: CrossAxisAlignment.center,
                     children: [
                       Text(
-                          '${notifier.completed() ? 'Rest' : notifier.step.actualReps}',
-                          style: TextStyle(fontSize: 64)),
+                        '${!notifier.started() || notifier.completed() ? 'Rest' : notifier.currentStep.actualReps}',
+                        style: TextStyle(fontSize: 64),
+                      ),
                       Stack(
                         alignment: AlignmentDirectional.center,
                         children: [
@@ -103,24 +106,24 @@ class AMRAPStepScreen extends StatelessWidget {
                     ],
                   ),
                 ),
-              ],
-            ),
-            ElevatedButton(
-              onPressed: () => getNextScreen(ctx),
-              child: Text(
-                getButtonText(ctx),
-                style: TextStyle(fontSize: 24.0),
               ),
+            ],
+          ),
+          ElevatedButton(
+            onPressed: () => getNextScreen(ctx),
+            child: Text(
+              getButtonText(ctx),
+              style: TextStyle(fontSize: 24.0),
             ),
-          ],
-        ),
+          ),
+        ],
       ),
     );
   }
 
   String getButtonText(BuildContext context) {
-    AMRAPChangeNotifier notifier =
-        Provider.of<AMRAPChangeNotifier>(context, listen: false);
+    AMRAPChangeNotifer notifier =
+        Provider.of<AMRAPChangeNotifer>(context, listen: false);
     if (notifier.completed()) {
       return 'Next Step';
     } else if (notifier.started()) {
@@ -131,15 +134,13 @@ class AMRAPStepScreen extends StatelessWidget {
   }
 
   void getNextScreen(BuildContext context) {
-    AMRAPChangeNotifier notifier =
-        Provider.of<AMRAPChangeNotifier>(context, listen: false);
-    if (notifier.completed()) {
-      nextStep();
-    } else if (notifier.started()) {
-      notifier.endTimer();
-    } else {
+    AMRAPChangeNotifer notifier =
+        Provider.of<AMRAPChangeNotifer>(context, listen: false);
+    if (!notifier.currentStep.isStarted) notifier.startStep();
+    else if (!notifier.currentStep.isCompleted) notifier.completeStep();
+    else {
       _controller.restart(duration: 60);
-      notifier.startTimer();
+      notifier.start();
     }
   }
 }
