@@ -3,9 +3,7 @@ import 'package:fitnet/models/workoutStepInstance.dart';
 import 'package:flutter/material.dart';
 
 class RepsForTimeStepScreen extends StatelessWidget {
-  final RepsForTimeStepInstance step;
-
-  RepsForTimeStepScreen({@required this.step});
+  RepsForTimeStepScreen();
 
   @override
   Widget build(BuildContext context) {
