@@ -1,4 +1,6 @@
+import 'package:fitnet/models/set.dart';
 import 'package:fitnet/models/workoutStepInstance.dart';
+import 'package:fitnet/shared/notifiers/instanceChangeNotifier.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
@@ -6,21 +8,24 @@ import 'package:provider/provider.dart';
 import '../../../shared/notifiers/setsChangeNotifier.dart';
 import 'currentSetDisplay/currentSetDisplay.dart';
 import 'setsDisplayRow/setsDisplayRow.dart';
+class Test {
+  List<Set> sets = [];
 
+  Set get currentSet => sets.firstWhere((set) => !set.isComplete());
+
+  Test();
+}
 class SetBasedStepScreen extends StatelessWidget {
-  final SetBasedStepInstance step;
-  final Function nextStep;
-
-  SetsChangeNotifier _notifier;
-
-  SetBasedStepScreen({@required this.step, @required this.nextStep}) {
-    _notifier = SetsChangeNotifier(sets: step.sets);
-  }
+  SetBasedStepScreen();
 
   @override
   Widget build(BuildContext context) {
-    return ChangeNotifierProvider<SetsChangeNotifier>.value(
-      value: _notifier,
+    return ChangeNotifierProxyProvider<InstanceChangeNotifier, SetsChangeNotifier>(
+      create: (ctx) => SetsChangeNotifier(parent: Provider.of<InstanceChangeNotifier>(ctx, listen: false)),
+      update: (_, instance, notifier) {
+        notifier.setParent(instance);
+        return notifier;
+      },
       builder: (_, __) => Column(
         children: [
           Expanded(flex: 5, child: SetsDisplayRow()),
@@ -34,7 +39,7 @@ class SetBasedStepScreen extends StatelessWidget {
                   builder: (ctx, notifier, __) => ElevatedButton(
                     onPressed: () => getNextScreen(ctx),
                     child: Text(
-                      getButtonText(ctx),
+                      getButtonText(notifier),
                       style: TextStyle(fontSize: 24.0),
                     ),
                   ),
@@ -47,26 +52,22 @@ class SetBasedStepScreen extends StatelessWidget {
     );
   }
 
-  String getButtonText(BuildContext context) {
-    SetsChangeNotifier setsNotifier =
-        Provider.of<SetsChangeNotifier>(context, listen: false);
-    if (setsNotifier.currentSet.isInProgress()) {
-      return setsNotifier.currentIndex == setsNotifier.sets.length - 1
-          ? 'Finish Step'
-          : 'Finish Set';
-    } else {
-      return 'Start Set';
+  String getButtonText(SetsChangeNotifier setsNotifier) {
+    if(setsNotifier.currentSet == null) {
+      return 'Finish Step';
     }
+
+    if (setsNotifier.currentSet.isInProgress()) {
+      return 'Finish Set';
+    }
+
+    return 'Start Set';
   }
 
   void getNextScreen(BuildContext context) {
-    SetsChangeNotifier setsNotifier =
+    SetsChangeNotifier sets =
         Provider.of<SetsChangeNotifier>(context, listen: false);
-    if (setsNotifier.currentSet.isInProgress()) {
-      setsNotifier.finishCurrentSet();
-      if (setsNotifier.currentIndex == setsNotifier.sets.length) nextStep();
-    } else {
-      setsNotifier.startNextSet();
-    }
+    if (sets.currentSet.isInProgress()) sets.completeSet();
+    else sets.startSet();
   }
 }
