@@ -1,6 +1,7 @@
 import 'package:fitnet/models/workoutInstance.dart';
 import 'package:fitnet/shared/completedInstanceRow/completedInstanceRow.dart';
 import 'package:fitnet/shared/inProgressInstanceRow/inProgressInstanceRow.dart';
+import 'package:fitnet/shared/notifiers/listChangeNotifier.dart';
 import 'package:fitnet/utils/customColors.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
@@ -20,10 +21,10 @@ class ViewWorkoutLatestInstance extends StatelessWidget {
           child: Row(
             mainAxisAlignment: MainAxisAlignment.spaceBetween,
             children: [
-              Selector<List<WorkoutInstance>, WorkoutInstance>(
-                selector: (_, instances) => instances[0],
+              Selector<ListChangeNotifier<WorkoutInstance>, WorkoutInstance>(
+                selector: (_, notifier) => notifier.list[0],
                 builder: (_, latestInstance, __) => Text(
-                    latestInstance.isCompleted()
+                    latestInstance.isCompleted
                         ? 'Last Workout:'
                         : 'In Progress Workout:',
                     style: Theme.of(context).textTheme.subtitle1),
@@ -47,9 +48,10 @@ class ViewWorkoutLatestInstance extends StatelessWidget {
             ),
             margin: EdgeInsets.fromLTRB(15, 0, 15, 15),
             padding: EdgeInsets.all(10),
-            child: Selector<List<WorkoutInstance>, WorkoutInstance>(
-                selector: (_, instances) => instances[0],
-                builder: (_, latestInstance, __) => latestInstance.isCompleted()
+            child: Selector<ListChangeNotifier<WorkoutInstance>, WorkoutInstance>(
+                selector: (_, notifier) => notifier.list[0],
+                shouldRebuild: (prev, next) => true,
+                builder: (_, latestInstance, __) => latestInstance.isCompleted
                     ? CompletedInstanceRow(
                         instance: latestInstance,
                         circleSize: 75,
