@@ -1,6 +1,8 @@
 import 'package:fitnet/models/workout.dart';
 import 'package:fitnet/models/workoutInstance.dart';
 import 'package:fitnet/services/workoutInstanceService.dart';
+import 'package:fitnet/shared/notifiers/listChangeNotifier.dart';
+import 'package:fitnet/shared/notifiers/workoutInstancesChangeNotifier.dart';
 import 'package:fitnet/utils/customColors.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
@@ -17,9 +19,20 @@ class WorkoutHistoryScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return StreamProvider<List<WorkoutInstance>>(
-      initialData: [],
-      create: (_) => instanceService.getInstancesStreamByWid(workout.wid),
+    return MultiProvider(
+      providers: [
+        FutureProvider<List<WorkoutInstance>>.value(
+          initialData: [],
+          value: instanceService.getInstancesForWorkout(workout.wid),
+        ),
+        ChangeNotifierProxyProvider<List<WorkoutInstance>, ListChangeNotifier<WorkoutInstance>>(
+          create: (_) => ListChangeNotifier<WorkoutInstance>(),
+          update: (_, instances, notifier) {
+            notifier.list = instances;
+            return notifier;
+          }
+        )
+      ],
       builder: (ctx, _) => Scaffold(
         appBar: AppBar(
           centerTitle: true,
