@@ -4,15 +4,15 @@ import 'package:flutter/material.dart';
 class CustomColors {
   CustomColors._();
 
-  static const Color blue = Colors.blue;
-  static const Color green = Colors.green;
+  static const MaterialColor blue = Colors.blue;
+  static const MaterialColor green = Colors.green;
   static const Color black = Colors.black;
   static const Color white = Colors.white;
-  static const Color yellow = Colors.yellow;
-  static const Color red = Colors.red;
+  static const MaterialColor yellow = Colors.yellow;
+  static const MaterialColor red = Colors.red;
 
   static const Color darkGrey = const Color(0xFF505050);
-  static const Color grey = Colors.grey;
+  static const MaterialColor grey = Colors.grey;
   static const Color lightGrey = const Color(0xFFDCDCDC);
 
   static getColorForCompletion(double percentage) {
